net: ethernet: ti: netcp_core: add ability to set number of tx channels
This allows to control number of tx channels within netccp core and
is supposed to be used with gbe. This support is completely dependent
on options in DT and is supposed to be not used if DT doesn't contain
appropriate tx queues. Initiall intention is to add k2g multichannel
support and then shaper configuration for every tx queue.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>