arm: boot: dts: keystone-k2g-netcp: add queues for tx pools
In multichannnel mode, for k2g SoC with CPSW-2U, every tx channel
has it's own pool, but number of general queues is 8, 4 of them are
used for rx, tx and other pools. According to documentation, it's
possible to add 7 additional queues to cover every tx channel,
leaving 4 in reserve, as it was before this change.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>